pyproject.toml
[project]
name = "portfolio-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "markdown>=3.10.1",
    "pydantic>=2.12.5",
    "pypdf>=6.6.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
    "unstructured>=0.18.27",
    "uvicorn>=0.40.0",
]
